Tidy memento refs (#154)
Tidy Memento references to be more consistent. Use forms:

"Memento specification [RFC7089]" when we refer to the spec in general
"Memento [RFC7089] section x.y" or just " [RFC7089] section x.y" when we refer to a specific section
"Memento" alone refers to a resource that encapsulates a prior state of the Original Resource (not the spec).